requirements.txt
Flask==2.2.2
psycopg2==2.9.3
python-dotenv==0.21.0
SQLAlchemy==1.4.41
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-WTF==1.0.1